requirements.txt
Django==1.9.1
django-bootstrap3==6.2.2
django-markdown==0.8.4
django-recaptcha==1.0.5
django-tables2==1.1.1
easy-thumbnails==2.3
Markdown==2.6.5
mysqlclient==1.3.7
Pillow==3.1.0
six==1.10.0
wheel==0.24.0
